add search engine

+++ app/Http/Controllers/TestimonyController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\testimony;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Storage;
 
 class TestimonyController extends Controller
@@ -23,7 +24,7 @@ class TestimonyController extends Controller
         $request->validate([
             'nama' => 'required',
             'komentar' => 'required',
-            'foto' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:4048',
+            'foto' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:8048',
         ]);
 
         $foto = Storage::putFile('public/testi',  $request->foto->path());
@@ -44,27 +45,30 @@ class TestimonyController extends Controller
     public function update(Request $request, $id)
     {
         // dd($request);
-        $data = ([
+        $validator = Validator::make($request->all(), [
             'nama' => 'required',
             'komentar' => 'required',
         ]);
         if($request->foto){
-            $data['foto'] = 'required|image|mimes:jpg,png,jpeg,gif,svg|max:4048';
+            $validator = Validator::make($request->all(), [
+                'foto' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:8048',
+            ]);
         }
-        $request->validate($data);
 
-        // update foto
-        $targetItem = testimony::where('id', $request->id)->first();
-        if($request->hasfile('foto')){
-            Storage::delete($targetItem->foto);
+        if ($validator->passes()) {
+            // update foto
+            $targetItem = testimony::where('id', $request->id)->first();
+                if($request->hasfile('foto')){
+                    Storage::delete($targetItem->foto);
 
-            $fotoutama = Storage::putFile('public/produk',  $request->foto->path());
-            $update['foto'] = $fotoutama;
-        }
+                    $fotoutama = Storage::putFile('public/produk',  $request->foto->path());
+                    $update['foto'] = $fotoutama;
+                }
             $update['nama'] = $request->nama;
             $update['komentar'] = $request->komentar;
             testimony::where('id', $id)->update($update);
-        // return redirect()->back();
+            // return redirect()->back();
+          }
     }
     public function delete($id)
     {
